date_me/requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
carehare==1.0.2
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.0
multidict==6.1.0
packaging==24.2
pamqp==3.3.0
pika==1.3.2
propcache==0.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
service-identity==24.2.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
yarl==1.18.3
zope.interface==7.2
